Moved components to individual files
