+++ app/addons/cors/components_old.js
@@ -1,0 +1,43 @@
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not
+// use this file except in compliance with the License. You may obtain a copy of
+// the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations under
+// the License.
+
+// import app from "../../app";
+// // import FauxtonAPI from "../../core/api";
+// import React from "react";
+// import Stores from "./stores";
+// import Actions from "./actions";
+
+
+
+// var validateOrigin = function (origin) {
+//   if (!Resources.validateCORSDomain(origin)) {
+//     FauxtonAPI.addNotification({
+//       msg: 'Please enter a valid domain, starting with http/https.',
+//       type: 'error',
+//       clear: true
+//     });
+
+//     return false;
+//   }
+
+//   return true;
+// };
+
+
+
+
+
+
+
+
+
+
